requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
mcp==0.9.1
pytest==8.3.0
aiohttp==3.11.0
pytest-asyncio==0.25.0
python-dotenv==1.0.0